[fdf] Block drivers from driver index
Driver index will not return a driver in MatchDriver and MatchDriverV1
FIDL response if the driver is marked as not allowed. Driver index
retrieves the value of the boot argument `devmgr.disabled-drivers` to
determine which drivers are not allowed. The value of
`devmgr.disabled-drivers` is a string that is a comma-separated list of
component URL's of drivers that are not allowed.

Driver test realm will create the `devmgr.disabled-drivers` boot
argument using the realm arguments passed into it on creation.

Bug: 92329
Change-Id: I10f62875015e76fc70d574d07f446ecd5933f320
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644842
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>