[fidl][cpp] Fix llcpp_interop_tests flake
The test makes a call with an infinitely old timeout. As a result, when
the request was invalid, it will not immediately fail: sometimes the
channel call returns early enough with the expected timeout error,
other times the server reads the invalid message first and closes the
channel.

Also adds an assertion to make sure any message we `Write` or `Call`
down a transport is a transactional message.

Fixed: 94400
Change-Id: Idf6838af09791b98598dcea02ff7c72055d6cbe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654303
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>