[artifactory] Upload md5-mismatched files for debugging
When artifactory detects that two different builds produced different
contents for a file that is supposed to be deduplicated, upload the file
to an "md5-mismatches" subdirectory of the build namespace directory to
aid in debugging. The other version of the file already exists in GCS so
we only need to upload the local version.

Bug: 78017
Change-Id: Ic0d2dff4db2074fc5707b9458ee874b6beb3a666
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655067
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>