[component][fuzzing] Add CorpusReaderClient
This CL adds an AsyncFidlClient class, and extends it into a
CorpusReaderClient. Additional test fixtures are also included to make
it easier to fake FIDL servers in async tests.

It does not add the CorpusReaderClient to the controller; that change is
saved for a subsequent CL.

Bug: 92490
Change-Id: I740984bfc9083c6fdc7d81615d583cf05bb1067b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642092
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>