[sdk][build] Fix merged_sdk archive copying
Previously merged_sdk was actually copying the stamp file from the gen
dir to the out dir instead of copying the actual archive.

Also fixed incorrectly using the current target's gen dir as the dep's
gen dir, which works when the target and all dependencies are in //sdk,
but not elsewhere.

Bug: b/221290511
Change-Id: I1ce95a4571e60fd16c471732f85622cb5f797ec3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655187
Reviewed-by: Misha Gridnev <gridman@google.com>
Commit-Queue: Sean Cuff <seancuff@google.com>