[fidl][hlcpp] Support table/union payloads
Adds support for top-level union/table payloads, like so:

```
protocol MyProtocol {
  MyMethod(table{ 1: a bool; } -> (union{ 1: a bool; 2: b int8; });
};
```

This change modifies the codegen for all CPP bindings, since it was much
easier to change the fidlgen_cpp schema and all of its consumers in the
same CL. It also proved difficult to do HLCPP without doing LLCPP as
well, and HLCPP needed to go first, since that is what
`fidl-compatibility-test` uses to define types used to specify the tests
in the compatibility_tests.cc test file. So, all CPP codegen is
implemented in this CL, with the following CLs marked `[llcpp]` and
'[cpp]` implement the compatibility tests servers only.

Test: fx regen-goldens; fx test fidl-compatibility-tests
Bug: 88343
Change-Id: I5c236f5eb64d2311d792295faee89a1f7d076173
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650811
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>