[test_manager] Allow dart/flutter runners and route required services
In order to run a CFv2-based Dart test, I developed a Rust realm_builder
shim that loads the required runner into the dart component's
environment, and eagerly launches the dart test component.

Because the Dart-based test requires loading a Dart or Flutter Runner
into the environment (and the Runner requires some additional privileges
and services not offered to hermetic tests) the test must run as a
`system-test` at this time.

Also note that it may not (yet) be possible for the test manager to
capture the exit status from the Dart test to the test manager. The exit
status is typically used to properly report tests passing vs. tests
failing. However, with the described Rust realm_builder workaround, test
manager does report the test failure if the test logs an error (which
the dart::test implementation currently does).

Change-Id: Ic839c68406c598c9dc4fc598bc214dd17ae3679f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652767
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>