[component][fuzzing] Add RunnerPtr
Shared pointers are easier and safer to use with asynchronous futures
and promises. This CL aliases std::shared_ptr<Runner> as RunnerPtr, and
updates call-sites to use it.

It also provide MakePtr as a way to create a RunnerPtr. Although not
needed in this CL, the Runner will soon want to access to the shared
ExecutorPtr. This CL includes the pointer as a constructor parameter
rather than updating all the Runner constructors again in a subsequent
CL.

Bug: 92490
Change-Id: I9eb402b3a1b0a53235afc579fa10a255c98367b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644534
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>