[size-check] Run size_check as part of the build
The legacy tool and the new tool are both running
as part of the build and the output are compared
with a tolerance.

Size checker and budget generation are moced to
build/images/assembly to break the dependency cycle.

check_size_depfile tool enumerates the manifests
files used by the size_check so that it can be used
as depfile.

Bug: 94824
Change-Id: Idef773818b33a5f4ec21328c46a9e3f7d4ad1bab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625562
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Pierre Labatut <plabatut@google.com>
Fuchsia-Auto-Submit: Pierre Labatut <plabatut@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>