[fxfs] Fallback to DeepCopy for Inspect snapshot
By default the Inspect runtime uses frozen mode for snapshotting an
Inspect VMO to avoid expensive copies. This is done by creating a child
VMO based on the backing Inspect VMO. In the event this operation fails,
the default fallback mode is to send a handle to the live Inspect VMO.

This can cause issues with the tree reader obtaining a consistent tree
snapshot under heavy filesystem load, so instead we send a deep copy of
the backing VMO to avoid this.

Make `frozen_fails_with` function in `TreeServerSendPreference` and
`send_vmo_preference` field in `TreeServerSettings` public.

Bug: 57330
Test: fxfs-fs-tests
Change-Id: Idf2bdd552f3072db46d043069b4260bd9b55c5d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652910
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>