[ffx][driver][dump] Display a separate node in the graph for each device with the same name
With the current implementation, devices with the same name are displayed as the same node in the resulting graph which is a bit confusing (eg. it looks like there is only one "hid-device" device with many parents.)

This implementation also makes it easier to customize each node with extra styles/info based on device info in the future.

Change-Id: Iaad1a15fbeb4f60f34789a39f01cfe143f972050
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655402
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Adam Chyb <achyb@google.com>