[check-licenses] Adding 'file' package
Simple file struct to hold more information about each file in the
fuchsia tree.

This also allows me to specify what type of file it is in the file
struct (e.g. a LICENSE file, a NOTICE file (and what type)), etc.
For me, I find that this sort of OO structure is easier to follow.

In the future I'd like to also use this struct to access the file
contents in a more efficient manner (only read once, keep the data cached)

Change-Id: I7d67c9c8928bf4cd44f7f78926b506652e88eb30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655384
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>