[microbenchmarks] Disable building the host version of the tests for Mac
Disable building fuchsia_microbenchmarks for Mac, because debugging
problems with running this on Mac has turned out to be a pain.

For example, mutex.cc depends on the C11 <threads.h>, which is not
provided by the Mac toolchain.  (It's not provided by the
currently-pinned Linux sysroot either, but it should be provided in
newer versions.)

This relands part of this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/647584

Bug: 93071
Change-Id: I80bcbd8fa8f7a7f2b5b46356404df77b6a8a8dd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655665
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>