[component][fuzzing] Make more Controller methods async
This CL modifies AddToCorpus, WriteDictionary, ReadDictionary, and
GetResults to make them use AsyncSocket methods.

As part of this change, the std::unique_ptr<Runner> held by the
Controller is changed to a more promise-friendly std::shared_ptr, which
is aliased as RunnerPtr.

The tests are unchanged.

Bug: 92490
Change-Id: If102319a57326b9a9cb9dd958d6743ef8372439a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644535
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>