[hypervisor][apic] Support NMI delivery mode
When a guest writes to the ICR register in the x2apic with an NMI
delivery mode (0x4) then we need to generate an IPI to the target vcpus.
For these interrupts the vector field is ignored because the NMI vector
is already well defined (2).

Bug: 93009
Change-Id: I363dc0676fbd8dbec41c7504f9794f56898edfe3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655646
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>