[build] Check rust commands for output-dir leaks
When using rewrapper --canonicalize_working_dir (new in 0.57),
verify that this transformation would be safe for caching,
i.e. be consistent across builds with different output dir (with all
other parameters matching).

For now, we check the command and the declared outputs for
leaks of the local output dir (quick).
Remote-produced outputs are not scanned (yet).

We could not simply wrap the rewrapper... command with output-scanner.sh
because rewrapper expects --output_files to contain the $build_subdir
(output directory).

--canonicalize_working_dir is not turned on yet (still finding issues).
Providing this check now will make it easier for others to help.

Bug: 94355
Bug: 92670
Change-Id: Ia719d71648475f9f5f555c74a8b56d69aa04d423
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655246
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>