[netstack3][ndp] Verify address conflict in test
Change the test to first figure out what temporary SLAAC address would
be generated, then assign it before it is generated. Add a counter that
gets incremented on conflict and is read in the test. Together, this
verifies that the test is actually causing a conflict, before checking
the behavior that should follow after.

Without this, it's possible for the code to be changed to not handle
conflicts correctly and for the test to still pass.

Bug: 21428

Change-Id: I53cb32dd44070bc2f06343c47d3ea004411480db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652042
Reviewed-by: Arthur Sfez <asfez@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>