[build][rbe] Suppress --canonicalize_working_dir.
For this target, the Rust command depends on the output-dir
as-is, and cannot be relocated to a different output-dir.
Mark it as such with the pseudoflag:
--remote-flag=--canonicalize_working_dir=false.

RBE accounts for this in the action hash, and will do the right thing
w.r.t. caching.

Bug: 94355
Change-Id: I880196a01815d49375b831208a69c98878074a70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655324
Reviewed-by: Andy Weiss <dragonbear@google.com>
Commit-Queue: David Fang <fangism@google.com>