[assembly] Tool providers can use paths to access tools.
With this change the zbi signing script can now be a tool, because the
tool providers can create tools by their path. This is important so that
we can log the usage of the zbi signing scripts.

Bug: 92883
Test: fx test assembly_tool
Change-Id: I5b63678b5261baa8f7e69e4cceb8e98f31dccec8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654264
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>