[assembly] Use the tool provider for zbi signing
This change adds the zbi signing script to the command log by running
the script via the ToolProvider interface.

We are doing this so that we can ensure that we do not break zbi signing
when we switch to create-system.

Bug: 92883
Test: fx test assembly
Change-Id: I6c230ab40cb0cd50507fe7dc35ba999fe05a25c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654267
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>