[fms] add device specific configuration to virtual device spec
This is a reland of fxr/646802 with a new check for whether or not
we should generate a physical or virtual device spec or both.

From fxr/646802, this commit allows the default window-size to be
overridden using a gni variable. In order to distinguish between
similar virtual device specs using the same board but different
window-size, adding a prefix var is required as well.

Bug: 88619
Change-Id: Icc19e0781b915d2d017ff36521d779c1b80d0a8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651724
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Commit-Queue: Omer Lev-ran <omerlevran@google.com>