[i915] Separate display controller device.
Previously, the root 'i915' device also served as the device the display
controller driver bound to.

In DFv2, this is not supported. This CL make the root 'i915' device into
a NON_BINDABLE device and exposes the display controller protocol through
a passthrough child.

Bug: 91608
Change-Id: Ic1600f8798ad6d69da14fa33a78c108dd7c92532
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650092
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>