[assembly] Strip both prefixes when comparing the command logs
When generating the zbi, a path to the kernel is provided, which is in
the legacy bundle, which uses the reference outdir. This means that when
stripping the paths for the comparison we need to strip both outdirs.

This is an example of the path used in both command logs:
  obj/build/images/zedboot/zedboot/legacy/kernel/kernel.zbi

Bug: 92883
Change-Id: I3598299baaec4719c60a4337ad17eb9b4c06e200
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654274
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>