[lib][zircon] Remove test for debug address
This feature is tested extensively in utest/debugger, and the getter
and setter implementations are done by a standard macro. The issue with
testing this is that the debug address of the process may be set by
other parts of the system, and the property is only allowed to be set
once.

Fixed: 94568

Change-Id: If5563987db36441408e49cf76034fe1cc47b1a53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656003
Reviewed-by: Dangyi Liu <dangyi@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>