[system-tests] Disable LGB->HEAD OTA tests
These tests verify that the last green build from some release builder
can OTA to the build under test. Generally this works well, but this
breaks down when we cut a release branch. For example, as of now in F6
we're paving the latest build from workstation, which is running F7,
then trying to OTA the build to F6, which is disallowed.

This disables these tests for now in the release branch in order to
fix the builds.

Bug: 94744
Change-Id: I775d15d19647a3ec9c9cd2add667c4560fc34958