[bt] Remove RfcommTest routes and define testonly manifests
Remove capability routes to `RfcommTest` so that it is not available in
production environments.

Instead, define testonly manifests with the capability routes. Users
will need to manually update core.cml to point to the testonly manifest
if needed. At this time, we do not want to go through the process of
making #bluetooth-core product variable.

Remove appmgr route to RfcommTest as it will be routed through the
testonly manifest.

Define a testonly group for Bluetooth core. This can be included in the
relevant product.gni file to set up a special BT realm for testing
scenarios.

Bug: 86390
Test: fx build && manually verify bluetooth-core starts up okay
Test: fx build with bt-init-testonly.cml and verify RfcommTest is
available

Change-Id: I484bfe482a89184e26a263ca202f16ae65d2d8d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653224
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>