[ffx][repo] Fix ffx repository package {list,show}
These broke because we were prepending "blobs/" to the paths when trying
to determine the package timestamp. This was missed due to a lack of
tests.

Along the way, this changes these functions to use the target length
entry, rather than the custom field "size", which isn't used in the
package resolver.

Fixed: 79915
Change-Id: Ifd371ab44f8b35988d4e4ce0d82116f6a6c79440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656663
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>