[driver_manager] Fix fuchsia.compat.LIBNAME property for DFv1
BindDevice() is called with an absolute path to a library, but the
fuchsia.compat.LIBNAME property should just be the library's name.

We fix this by adjusting driver_host to pass just the libname to
driver_manager, and then driver_manager uses that to match the device.

Bug: 91608
Change-Id: I8175191255e07bcc227bffe57a57f116cf882296
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654884
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>