[ffx][preflight] Gather CPU features from cpuinfo
On Discord, roddux#5128 (an external contributor) pointed out that
`lscpu` makes use of gettext, and that our regex assumes that the person
running `ffx platform preflight` is using a system configured with
`en_US` i18n/l10n settings.

This change switches from using output of the `lscpu` command to looking
for either the `vmx` (for VT-x) or `svm` (for AMD-V) flags in its
output. We may like to consider writing an API for consuming
`/proc/cpuinfo` output, but that's outside the scope of this change.

Test: `lscpu` tests modified to check cpuinfo from real-world devices.
Change-Id: I0b4b13e13d3e7ffac90415c13a4ed3717720c269
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658282
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Devon O'Dell <dhobsd@google.com>