[rust_crates] Disable global and add local OWNERS
This disables global OWNERS due to the need to enforce certain processes
that global OWNERS may not be familiar. There have been a couple recent
examples of this. Since this may increase the load on the (already
small) set of local OWNERS I'm also adding core Rust on Fuchsia team
members to it.

Change-Id: I65706f5a72a21ef19588366f7046a6c28d58b5fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656666
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>