[audio][drivers][hda] Add MISCDBCGE bit clearing
It is recommended by HW that the CGCTL register MISCDBCGE bit is
deasserted when a reset of the HDA controller HW is performed.
This bit controls the "Clock Gating Control"
"Backbone Dynamic Clock Gating".

This was found investigating very rare boot HDA initialization
failures in fxrev.dev/84037, but it is not known to fix these.

Test: intel-hda-controller-test

Change-Id: Ic05c2c83dbb2d65c7b80107ad8999ec59dc54860
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658377
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>