[network] Introduce todo_unused macro
The macro serves as a code manifestation of the cultural agreement to
balance two needs that are often in opposition:
- keep CLs small
- avoid introducing code that is not used outside of tests

The macro enforces that a bug number be informed to track the technical
debt with urgency, and annotates statements with `cfg(test)` which
require revisitation organically when the code becomes load-bearing.

This does not address the unfortunate virality of `cfg(test)`, but we
take the advantages above to be sufficient.

Change-Id: I0b70f0689a75a78e5c6d2343e4a369a3b8034e03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657592
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>