[audio][test] Handle single-frequency tests
A frequency specified in ".single_frequency_to_test" should
ideally be in kReferenceFrequencies[] so we have FR and SiNAD
results for it. If it is not, display a warning but continue,
actually testing at the nearest frequency that IS in our set
(rounding down); we do this (rather than CHECK) to make single-
frequency spot-checking easier during ad-hoc investigations.

Other adjustments in how we display single-frequency results.

fx test audio-core-fidelity-test                        \
        audio-core-google-pipeline-fidelity-test-nelson

Change-Id: Ic6fcb44b331acc0cc3413f52ed035189d8274abf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658584
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>