[fidl][cpp] Client API uses fidl::Result
This CL implements the client part of integrating with |fitx::result|
and folding errors in the unified bindings:

- No more client response callbacks.
- Callback takes |fidl::Result<Method>&|.

The only difference from the design doc is that we now inherit from
|fitx::result| per advice from eieio@.

The two trait-like structs that provide information about a method are
merged into |fidl::internal::NaturalMethodTypes<Method>|.

The compatibility test got more verbose for servers sending errors,
because the server API still expect a union domain object in a struct.
That would be addressed in the next CL where we switch server API to use
fitx::result.

Bug: 90111
Change-Id: I96b102cd8d823dd25f25a7ba882c177e7cde1dbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658594
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>