[fint] Stop building system_snapshot explicitly
The infrastructure initially explicitly built the system snapshot so
that the size data it contained could be uploaded to GCS. However, no
code accesses the file via GCS and we instead use a different set of
size-tracking infrastructure, so we can stop uploading it from the
recipes (fxrev.dev/658882) at which point fint doesn't need to build it
anymore.

This change is safe to land before the corresponding recipe change
because the recipes will skip uploading the file if it doesn't exist on
disk.

Change-Id: I182fdc2257d12709cc2741354a89b2485b3ca6e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658748
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>