[starnix][binder] Implement ref-counting commands
These commands allow a binder process to increment/decrement the
weak/strong reference count of a binder object in a remote process.

In the future, the binder driver will keep its own ref count of objects
and only send a command to a binder process for the first and last
reference.

Change-Id: Ief46edabf4f1cc8e00ef413f0f5acef7bfeff893
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656566
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>