[netemul] Remove InterfaceOptions.enable_dhcp
Remove the option to start a DHCP client on the interface being
installed in a netemul runner test. This functionality is mostly unused
in existing runner tests, and is only used in a test that relies on the
netemul guest functionality, which will likely not be fulfilled by the
runner in CFv2.

Test: fx test netemul-runner-unit-tests
Change-Id: Iaaa9a1a53d544aea05aa20d8599906a5ff0470be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659185
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>