[bt][gap] Start asynchronously with local name
The local device's name is not guaranteed to be available in a timely
fashion - the NameProvider.GetHostName doesn't return until an ethernet
address is assigned.   Use the default host name until it returns.

If the default host name is overridden before the local name is set,
then prefer the name that was provided via
fuchsia.bluetooth.sys.Access.SetName

Bug: 94662
Test: fixed tests, added tests for new set_name functionality
Change-Id: Iaff19181e7e8ab419bb0a71b9fe4ef7558c3621f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652053
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>