[mixer_service] Add a sketch of mixer objects
This CL adds a sketch of two kinds of mixer objects, Thread and
PipelineStage. There are two implementations of Thread, one for the
"detached" thread and one for the mixer threads. See here for more
discussion:
//src/media/audio/mixer_service/docs/execution_model.md

This is just a sketch with not much functionality. Future CLs will flesh
out this directory.

Bug: 87651
Change-Id: Ifeb9ebda8d118cfd2e2d05b995521bde610d026b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649810
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>