[audio] Don't tune unstarted PIDs
PIDs use history for part of their feedback; start-time is
relevant. For most audio clocks, this is set when the stream
starts, when we discover that the source and dest clocks are
not identical. However, this doesn't happen automatically if
for example a client submits a custom clock that is a clone
of CLOCK_MONOTONIC, then subsequently rate-adjusts it.

The problem is not solved by immediately starting all PIDs
when their associated AudioClock is created. Any significant
duration from clock start to first tuning causes the PID to
overreact to the initial position error, because of how a
PID's Integral factor uses time-since-last-tuning.

To solve this, we now track a PID's most recent tune-time as
std::optional. If TuneForError() is called before Start(), we
set tune_time and current_error (to prepare for subsequent
TuneForError calls) but do NOT then generate a feedback
response based on that first tuning.

This defect is only surfaced by client-submitted clocks, and
the impact is exaggerated if a client rate-adjusts the clock
by a significant amount well before the stream starts (or if
for any other reason, initial position error is very large).

In practice, this combination of factors only occurs when a
Fuchsia device is used as a Bluetooth destination. In these
cases, synchronization logic in multizone-leader (the clock
owner) may continue to take a number of minutes to slowly
"unwind" the initial position error, doing this in a very
controlled way. After this fix, the Fuchsia device smoothly
and closely follows as it does so.

Fixed: b/223689874

Change-Id: Ic33eb5977204f0fc1bf474af56c8becd7e08e19d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658908
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>