[emu] Include QEMU prebuilt in the SDK
We've been expanding the feature set for emulation for some time, and
part of the new functionality is the ability to launch a more recent
version of QEMU than the branch they forked for AEMU. Currently, the
kernel team tests with 5.0, but they're actively upgrading to 6.0 for
base QEMU. AEMU is already available as a prebuilt in the SDK, but QEMU
is not.

This CL adds QEMU to the set of host tools, while importing the prebuilt
artifacts for SDK inclusion. This enables OOT users to run the QEMU5
emulator directly from the SDK as it's delivered, matching the
functionality for the AEMU backend.

Change-Id: If81eef6e4af98cd4141ceaae84aaa5009a5b5692
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657486
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>