[camera][imx227] Remove extra sensor power cycling
This change removes unnecessary power cycling between stopping
streaming and setting the mode when enabling new streams.

Removing the extra power cycling eliminates a period of time when
another piece of software might fail to check the on the sensor status
due it being powered down and unresponsive to I2C transactions.

It also leaves the power rails in a more stable state which can be
helpful for checking the amount of noise on the rails.

Bug: fxb/94060

Tested by running the daily camera test suite, which includes the unit
tests, camera-gym, and real-world use cases in various muting states.

Change-Id: Ibc0c520fd6a1b52933bff69009eb79c5937749fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659227
Reviewed-by: Ernest Hua <ernesthua@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>