[fidl][cpp] Cancellable callbacks in Zircon IPC
Introduces `Then` and `ThenExactlyOnce` which can be chained after a
two-way call to attach a continuation.

`Then` passivates the callback if the client object is destroyed.

`ThenExactlyOnce` calls it exactly once.

The next steps are to migrate user code and also update the driver and
natural messaging APIs to use this mechanism.

Bug: 94402

Change-Id: Ia63cba0967d7663ddbb925fdd248c6474bac9dd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650825
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>