[blobfs] Remove decompressor sandbox from unit tests.
This adds a tiny threaded version of the decompressor sandbox meant for
use within tests. This makes it so that we don't need to additionally
include the decompressor sandbox component for tests. It is easier to
add 2-3 lines of boilerplate than to juggle around the component
manifests for all the tests that need it outside of blobfs.

Bug: 62177

Change-Id: Id03f4f7552702aad3a900e4f133623f6839256fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630284
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>