[camera][gym] Add frame arrival deadline
For Thermal testing, it is important to fail the test run if there
are indications of catastrophic failures such as no frames arriving.

This CL adds a deadline for frame arrival such a new deadline is
established at the beginning of each stream and at each subsequent
frame. The tolerance for lateness is a constant 10 seconds for all
streams (regardless of the frame rate).

When this deadline is tripped (for any frame), camera-gym will
assert and crash out.

The thermal test rig will detect this failure and fail accordingly.

Bug: b/219806158
Test: Run thermal test normally.
      Run thermal test with forced sensor reset (test should fail).
Change-Id: Iba62b047f09e7dd0bc1f87acc0f01d529387fb90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656783
Reviewed-by: Matt Sandy <msandy@google.com>
Fuchsia-Auto-Submit: Ernest Hua <ernesthua@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>